Exer6 - Parte 2

+++ TrabalhoN2CG/src/cg/exer6/Ponto2D.java
@@ -2,28 +2,28 @@ package cg.exer6;
 
 public class Ponto2D {
 	
-	private int x;
-	private int y;
+	private float x;
+	private float y;
 	
-	public Ponto2D(int x, int y) {
+	public Ponto2D(float x, float y) {
 		super();
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
